Safari-611.1.10/LayoutTests/fast/forms/file/file-input-user-selection-events-expected.txt

OPEN FILE PANEL
This tests the events dispatched on user selection

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


on input
PASS inputDispatched is false
PASS changeDispatched is false
on change
PASS inputDispatched is true
PASS changeDispatched is false
PASS successfullyParsed is true

TEST COMPLETE
